cli/error.go: split a file in two
Release note: None
