+++ pkg/cli/error.go
@@ -24,7 +24,6 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/server"
 	"github.com/cockroachdb/cockroach/pkg/sql/pgwire/pgcode"
 	"github.com/cockroachdb/cockroach/pkg/util/grpcutil"
-	"github.com/cockroachdb/cockroach/pkg/util/log"
 	"github.com/cockroachdb/cockroach/pkg/util/netutil"
 	"github.com/cockroachdb/errors"
 	"github.com/lib/pq"
@@ -235,20 +234,4 @@ func MaybeDecorateGRPCError(
 		// Nothing we can special case, just return what we have.
 		return err
 	}
-}
-
-// maybeShoutError calls log.Shout on errors, better surfacing problems to the user.
-func maybeShoutError(
-	wrapped func(*cobra.Command, []string) error,
-) func(*cobra.Command, []string) error {
-	return func(cmd *cobra.Command, args []string) error {
-		err := wrapped(cmd, args)
-		return CheckAndMaybeShout(err)
-	}
-}
-
-// CheckAndMaybeShout shouts the error, if non-nil to the OPS logging
-// channel.
-func CheckAndMaybeShout(err error) error {
-	return clierror.CheckAndMaybeLog(err, log.Ops.Shoutf)
 }
